% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prepare_data.R
\name{fix_bills}
\alias{fix_bills}
\title{Function to take a data set, a legislator name, a majority party, and find bills to fix positions}
\usage{
fix_bills(legislator = NULL, party = NULL, vote_data = NULL,
  legislature = NULL)
}
\arguments{
\item{legislator}{Character string of the name of the legislator to use as a reference (ideal point fixed at 0)}

\item{party}{Character string of the majority party in the legislature (will be used to determine reference legislator's ideal points on bills)}

\item{vote_data}{Data frame with all of the votes for the particular legislature of analysis}

\item{party_data}{Data frame with all of the party and demographic information for legislators in both legislatures}
}
\description{
Function to take a data set, a legislator name, a majority party, and find bills to fix positions
}

