% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prepare_data.R
\name{fix_bills_discrim}
\alias{fix_bills_discrim}
\title{Function to take a roll call data set and the names of opposition and governing parties.
The opposition party should be as extreme in opposition as possible.
This function helps to achieve identification by constraining discrimination parameters.
For bills on which the opposition votes in majority, discrimination is constrained to be
positive.
For bills on which the government votes in majority, discrimination is constrained to be
negative.
For other bills, discrimination is free to float.}
\usage{
fix_bills_discrim(opp = NULL, gov = NULL, vote_data = NULL,
  legislature = NULL, to_run = NULL, use_nas = NULL)
}
\arguments{
\item{opp}{Character, name of opposition party to filter data}

\item{gov}{Character, name of governing party to filter data}

\item{vote_data}{Voting data in list form, with each element of list equal to a legislature}

\item{legislature}{The specific legislature in the list to choose}
}
\description{
Function to take a roll call data set and the names of opposition and governing parties.
The opposition party should be as extreme in opposition as possible.
This function helps to achieve identification by constraining discrimination parameters.
For bills on which the opposition votes in majority, discrimination is constrained to be
positive.
For bills on which the government votes in majority, discrimination is constrained to be
negative.
For other bills, discrimination is free to float.
}

