% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_data.R
\name{icc_ordinal}
\alias{icc_ordinal}
\title{Function to return item characteristic curves for ordinal response data (partial credit model)}
\usage{
icc_ordinal(steps = NULL, summary_data = NULL, bill_labels = NULL,
  param = NULL)
}
\description{
Function to return item characteristic curves for ordinal response data (partial credit model)
}

