% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_data.R
\name{posterior_predict}
\alias{posterior_predict}
\title{Function to calculate posterior predictive values of Y
Currently only works for split absence with discrimination parameters.}
\usage{
posterior_predict(summary_data = NULL, y = NULL, ll = NULL, bb = NULL,
  num_iters = 100)
}
\description{
Function to calculate posterior predictive values of Y
Currently only works for split absence with discrimination parameters.
}

