% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulate_data.R
\name{simulate_cleaned}
\alias{simulate_cleaned}
\title{A function taken from Daniel Furr's edstan package for the Rating Scale Model}
\usage{
simulate_cleaned(num_legis = 50, num_bills = 100, num_cuts = 3,
  legislature = NULL, parties = 3)
}
\description{
A function taken from Daniel Furr's edstan package for the Rating Scale Model
}

