% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prepare_data.R
\name{split_absences}
\alias{split_absences}
\title{Accepting already cleaned data, this function will split the outcome variable into a no --> abstain --> yes (1,2,3) component
And a absent/party votes no --> absent/party split --> absent/party votes yes (4,5,6)
The intention is to enable separate modeling of these two types of responses to legislative votes}
\usage{
split_absences(cleaned = NULL, to_fix = NULL)
}
\description{
Accepting already cleaned data, this function will split the outcome variable into a no --> abstain --> yes (1,2,3) component
And a absent/party votes no --> absent/party split --> absent/party votes yes (4,5,6)
The intention is to enable separate modeling of these two types of responses to legislative votes
}

